% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{assign_clusters}
\alias{assign_clusters}
\title{Assign the Clusters}
\usage{
assign_clusters(dist_mat, medoids)
}
\arguments{
\item{X}{Matrix n x d}

\item{centers}{Matrix K x d}
}
\value{
Integer Vector of cluster assignments
}
\description{
Assign the Clusters
}
