% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Compute euclidean distance}
\usage{
euclidean_distance(x, centers)
}
\arguments{
\item{x}{Numeric Vector}

\item{centers}{Numeric Vector}
}
\value{
Numeric euclidean distances between x and centers as vector
}
\description{
Compute euclidean distance
}
