% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{has_converged}
\alias{has_converged}
\title{Test if converged}
\usage{
has_converged(old_centers, new_centers, tol = 1e-06)
}
\arguments{
\item{old_centers}{previous centers}

\item{new_centers}{current centers}

\item{tol}{Tolerance for converged}
}
\value{
Logical, true if converged
}
\description{
Test if converged
}
