% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{K-Means clustering algorithm}
\usage{
k_means(X, K, max_iter = 100, tol = 1e-06, nstart = 10)
}
\arguments{
\item{X}{Numeric matrix or data.frame containing the dataset of size n x d}

\item{K}{Number of clusters}

\item{max_iter}{Maximum number of iterations (default 100)}

\item{tol}{Convergence tolerance (default 1e-6)}

\item{nstart}{Number of random starts (default 10)}
}
\value{
A list with following elements:
\item{centers}{Matrix of cluster centers (K x d)}
\item{clusters}{Vector of cluster assignments for each observation}
\item{iter}{Number of iterations until convergence}
\item{tot_withinss}{Total within-cluster sum of squares}
}
\description{
This is an implementation of the k-means clustering algorithm. (see Chapter 9.1, Stefan Richter - Statistisches und maschinelles Lernen)
It partitions a numeric data set into K clusters by iteratively assigning points to the nearest cluster center and updating the centers until convergence.
It supports multiple random starts to find a better local minimum.
}
\examples{
set.seed(1)
X <- rbind(matrix(rnorm(10*2, mean = 0), ncol = 2),
           matrix(rnorm(10*2, mean = 5), ncol = 2))
           
res <- k_means(X, K = 2)
res$centers
res$clusters

#Visualize clusters and centers
plot(X, col = res$clusters, pch = 19)
points(res$centers, col = 1:2, pch = 8)

}
