% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{update_centers}
\alias{update_centers}
\title{Update the Centers}
\usage{
update_centers(X, clusters, K)
}
\arguments{
\item{X}{Matrix n x d}

\item{clusters}{Integer Vector of Cluster assignments}

\item{K}{Number of Clusters}
}
\value{
Matrix K x d with updated Cluster Centers
}
\description{
Update the Centers
}
