% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{k-means algorithm}
\usage{
k_means(X, K, max_iter = 100)
}
\arguments{
\item{X}{Numeric matrix or data frame of size n x d}

\item{K}{Number of Clusters}

\item{max_iter}{Maximum number of iterations}
}
\value{
A list with cluster assignments and cluster centers
}
\description{
k-means algorithm
}
